backend_py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
motor==3.6.0
redis==5.0.8
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
pydantic==2.9.2


